test(dnsbl): split the #2371 pipeline-ordering TestCase into its own file
PHPUnit only discovers ONE TestCase per file: the second class in
PslFeedPolicySeedTest.php (TestFreshInstallPipelineOrdering, including the
sibling-field grandfather-mode regression guard) silently never ran under any
invocation. Move it to PslFeedPolicyPipelineOrderingTest.php (class name ==
file name, matching repo convention).

Also extract the install-time seed itself into pfb_install_psl_feed_policy_seed()
(pfblockerng.inc), a DI seam mirroring pfb_install_registry_writeback(), and have
pfblockerng_install.inc call it. The prior inline reimplementation inside the test
meant a mutation to the REAL installer file went undetected even after the file
split; the test now calls the production seam directly, plus a source-scan wiring
proof (mirrors InstallPrePassWriteOrderTest) and a DI-based ordering unit test.
Verified: dropping the icann_policy write from the real seam now fails 2 tests;
restored, green again.
